feat(tests): add zone intrusion + poketenashi tests, create 6 missing test datasets, fix stale paths
- test_p10_zone_intrusion.py: 11 tests (geometry, PiP, dataclasses, full detect+draw)
- test_p10_poketenashi.py: 19 tests (all 4 pose rule detectors + RuleResult)
- create_test_datasets.py: generates test_ppe_100, test_shoes_detection_100,
  test_fire_detection_100, test_phone_detection_100, test_fall_detection_100,
  test_segmentation from existing raw datasets (80 train + 20 val each)
- Fix stale fire dataset path in test_p00_data_prep.py (data/data/train → data/train)
- Fix fixtures checkpoint path (07_training → 08_training)
- Fix float32 crash in zone_intrusion.py draw() (cv2.polylines needs int32)
- Update 05_data_phone.yaml and 05_data_segmentation.yaml for new datasets

+++ features/access-zone_intrusion/code/zone_intrusion.py
@@ -131,7 +131,7 @@ class ZoneIntrusionDetector:
         vis = image_bgr.copy()
 
         for zone in self._zones:
-            poly_px = _poly_to_pixel(zone["polygon"], w, h)
+            poly_px = _poly_to_pixel(zone["polygon"], w, h).astype(np.int32)
             cv2.polylines(vis, [poly_px], True, (0, 255, 255), 2)
             overlay = vis.copy()
             cv2.fillPoly(overlay, [poly_px], (0, 255, 255))
